pyproject.toml
[tool.poetry]
name = "Bilingual-rag"
version = "0.1.0"
description = "Bilingual RAG Chatbot MVP"
authors = ["Hector Alamo <halamo24@gmail.com>"]
readme = "README.md"
packages = [{include="api"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
pydantic = "^2.7.0"
uvicorn = "^0.30.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.34"
pgvector = "^0.2.5"
redis = "^5.0.7"
httpx = "^0.27.0"
tiktoken = "^0.7.0"
nltk = "^3.9.1"
beautifulsoup4 = "^4.12.3"
sentencepiece = "^0.2.0"
transformers = "^4.44.0"
torch = "^2.3.0"
numpy = "^1.26.4"
orjson = "^3.10.7"
structlog = "^24.1.0"
pytest = "^8.2.0"
ragas = "^0.1.9"
"pdfminer.six" = "^20231228"
trafilatura = "^2.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
pytest-asyncio = "^0.23.8"
deepeval = "^1.0.0"

[tool.ruff]
line-length = 100
